refactor: SmsController AuthenticationApiController로 통합
- 회원가입/로그인 관련 기능 하나의 클래스로 통합
