+++ src/main/java/com/swifty/bank/server/api/controller/dto/auth/response/CheckVerificationCodeResponse.java
@@ -1,7 +1,6 @@
-package com.swifty.bank.server.api.controller.dto.sms.response;
+package com.swifty.bank.server.api.controller.dto.auth.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
