+++ src/main/java/com/swifty/bank/server/api/controller/dto/auth/response/SendVerificationCodeResponse.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.api.controller.dto.sms.response;
+package com.swifty.bank.server.api.controller.dto.auth.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
