+++ src/main/java/com/swifty/bank/server/api/service/SmsService.java
@@ -1,11 +1,11 @@
 package com.swifty.bank.server.api.service;
 
-import com.swifty.bank.server.api.controller.dto.sms.request.CheckVerificationCodeRequest;
-import com.swifty.bank.server.api.controller.dto.sms.request.SendVerificationCodeRequest;
-import com.swifty.bank.server.api.controller.dto.sms.request.StealVerificationCodeRequest;
-import com.swifty.bank.server.api.controller.dto.sms.response.CheckVerificationCodeResponse;
-import com.swifty.bank.server.api.controller.dto.sms.response.SendVerificationCodeResponse;
-import com.swifty.bank.server.api.controller.dto.sms.response.StealVerificationCodeResponse;
+import com.swifty.bank.server.api.controller.dto.auth.request.CheckVerificationCodeRequest;
+import com.swifty.bank.server.api.controller.dto.auth.request.SendVerificationCodeRequest;
+import com.swifty.bank.server.api.controller.dto.auth.request.StealVerificationCodeRequest;
+import com.swifty.bank.server.api.controller.dto.auth.response.CheckVerificationCodeResponse;
+import com.swifty.bank.server.api.controller.dto.auth.response.SendVerificationCodeResponse;
+import com.swifty.bank.server.api.controller.dto.auth.response.StealVerificationCodeResponse;
 
 public interface SmsService {
     StealVerificationCodeResponse stealVerificationCode(StealVerificationCodeRequest stealVerificationCodeRequest);
