+++ src/test/java/com/swifty/bank/server/api/service/SmsServiceTest.java
@@ -1,17 +1,17 @@
 package com.swifty.bank.server.api.service;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.swifty.bank.server.api.ConfigureContainer;
-import com.swifty.bank.server.api.controller.dto.sms.request.CheckVerificationCodeRequest;
-import com.swifty.bank.server.api.controller.dto.sms.request.StealVerificationCodeRequest;
-import com.swifty.bank.server.api.controller.dto.sms.response.StealVerificationCodeResponse;
+import com.swifty.bank.server.api.controller.dto.auth.request.CheckVerificationCodeRequest;
+import com.swifty.bank.server.api.controller.dto.auth.request.StealVerificationCodeRequest;
+import com.swifty.bank.server.api.controller.dto.auth.response.StealVerificationCodeResponse;
 import com.swifty.bank.server.core.common.redis.service.OtpRedisService;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.assertj.core.api.Assertions.*;
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @RequiredArgsConstructor
